+++ CHANGES.rst
@@ -6,6 +6,7 @@
 ========================
 
 - New ``compress_to`` methods for easy gzip compression with tarfile.
+- Added patterns support to ``Dir.files`` and ``Dir.subdirs`` methods.
 
 0.1.0 (2013-07-01)
 ==================