+++ CHANGES.rst
@@ -7,8 +7,9 @@
 
 - New ``compress_to`` methods for easy gzip compression with tarfile.
 - Added patterns support to ``Dir.files`` and ``Dir.subdirs`` methods.
-- Now ``Dir.files`` and ``Dir.subdirs`` methods return sorted list
+- Now ``Dir.files`` and ``Dir.subdirs`` methods return sorted list.
  allowing custom sort options and abspath args that allow to choose between relative/absolute path.
+- Added Dir.size method to recursively compute directory size.
 
 0.1.0 (2013-07-01)
 ==================