+++ CHANGES.rst
@@ -2,8 +2,8 @@
  Dirtools Changelog
 ====================
 
-0.1.1 (not released yet)
-========================
+0.1.1 (2013-12-12)
+==================
 
 - New ``compress_to`` methods for easy gzip compression with tarfile.
 - Added patterns support to ``Dir.files`` and ``Dir.subdirs`` methods.
@@ -11,12 +11,14 @@
  allowing custom sort options and abspath args that allow to choose between relative/absolute path.
 - Added Dir.size method to recursively compute directory size.
 - New ``DirState`` and ``compute_diff`` method.
-- Now skipping symlinks
+- Now skipping symlinks.
+
 
 0.1.0 (2013-07-01)
 ==================
 
 - v0.1.0 on pypi.
+
 
 0.0.0 (2013-06-17)
 ==================